font2h: a tool to convert a .ttf file to a .h file
Added a tool to convert TrueType fonts to a .h file. It uses Freetype to
render the characters and stores them in an array of 16-bit values.

Per character it stores a few metrics (width, height, top, left & advance)
and its pixels. Pixels are stored as a 4-bit value representing their
transparency. Only the pixels within the bounding box of a character are
stored.

Next to the array with the character metrics and pixel data, it also
generates a lookup array with indices to find a character back in the data
array.

Finally it adds a struct with references to these arrays and some
additional info, e.g. font height.

By default it renders all printable ascii characters: from 32 (space) to
126 (~), but it has a commandline argument to specificy which characters
it should render. This allows for rendering only specific characters of
a font to save space e.g. only the uppercase characters or only the
digits.

It currently doesn't support unicode characters...

Example:

    font2h -n opensans -s 12 OpenSans-Regular.ttf > font_opensans_12.h

Note: an additional package is required for building this tool:
libfreetype6-dev. Please make sure to rebuild the docker container
with the docker/build.sh script before compilng this!